fixup! Rework Flare on Windows to use all Winsock2 methods with OVERLAPPED mode

+++ src/Flare.h
@@ -53,7 +53,7 @@ private:
 #ifndef _MSC_VER
     Pipe pipe;
 #else
-    int sendfd, recvfd;
+    unsigned long long sendfd, recvfd;
 #endif
 };
